backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
deep-translator
Flask==3.1.2
Flask-Admin==2.0.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy
flatbuffers==25.9.23
frozenlist==1.8.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.52.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11
h2
h5py==3.15.1
hpack==3.0.0
hstspreload==2025.1.1
httpcore
httplib2==0.31.0
httpx>=0.28.1
hyperframe==5.2.0
idna
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.0
legacy-cgi==2.6.4
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.0
namex==0.1.0
numpy==2.3.5
openai==0.28.1
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL
pyparsing==3.2.5
python-dotenv==1.2.1
requests==2.32.5
rfc3986==1.5.0
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
WTForms==3.2.1
yarl==1.22.0
